Make PinIsCategory check for no interface
If the pin has no interface, then it's a custom made demuxer pin.  You
cannot query a category from a custom demuxer pin, so it'll be the
desired pin if this is used on a demuxer filter.  So if there is no
interface for IKsPropertySet, then assume it's the correct pin.