Add DirectConnectFilters helper function
This function attempts to directly connect the output pins of an output
filter to the input pins of an input filter automatically.  If any pins
were connected by the function, it returns true, otherwise false.

Does not use smart connect -- this function is meant to be used to
connect a crossbar filter to the main device filter.