Add support for FORMAT_VideoInfo2
FORMAT_VideoInfo2 is a commonly used format (whether I like it or not).
However, the foolish developers at microsoft decided to not make
VIDEOINFOHEADER2 compatible with VIDEOINFOHEADER for whatever reason, so
the only way to get the bitmap info header is by checking explicitly for
FORMAT_VideoInfo2 each time.

GetBitmapInfo must be used when checking for a valid media type to
ensure that it'll return the proper bitmap info header.