Add ShutdownGraph() function.
This function allows you to safely shut down the graph within a specific
thread that DirectShow may be used on.

I realized this function was needed when I started having issues trying
to use DirectShow in multiple threads.  You just can't do it; DirectShow
really must be contained to a single thread if you don't want to have
strange and unpredictable behavior.