Do not fail QueryAccept if running
IPin::QueryAccept is the function that's called to allow output filters
to dynamically change media types during the stream.  We want to allow
changing of media types during the stream, so this function must succeed
when called during in the middle of a stream to be able to do so.