Allow dynamic media type changes
When the media type dynamically changes on a pin, the next IMediaSample
sent will store the new media type used for the audio/video.

Make it so that the audio/video data callbacks include the current
AudioConfig/VideoConfig in order to allow the user of the library to
know when that has occurred.