Implement handling of encoded data
The microsoft demux filter always splits up encoded data in to sets of
8192-sized IMediaSample packets.  When the packets do not have a
timestamp associated with them, they are part of the previous set of
IMediaSample packets.  Otherwise, it signifies the start of a new set.
There are other ways to detect sets but this is a fairly simple means.

Instead of 'leaving this to the decoder', I decided to just combine the
packets in libdshowcapture before sending them out, to make using those
packets a bit more simple.