Fix HD-PVR Rocket support
Figured out how to get the rocket working properly, finally.  It's
somewhat of a pain.

Basically, the rocket defaults to this weird 'standby' mode.  While in
this mode, the rocket will not stream any data, regardless of filter
connections.  I figured out how to go in to 'capture' mode via
monitoring the DirectShow API when it's being used by other
applications.  I discovered one key call that enabled it:
calling IPropertyKeySet::Set on the encoder with some specific
parameters.  I managed to duplicate the functionality myself by calling
the same function with the same parameters, and presto, the device can
now be activated/deactivated.