Fix a bug with an error message
Passing an std::string as a printf parameter won't work.  You have to
call std::string::c_str() to be able to properly use it in this case,
although printf-type functions do not have any type checking for this
(at least on terrible microsoft compilers), so it's easy to
inadvertently pass it as std::string.