Allow empty names when finding filters
Most devices have two primary values to determine which device it is,
its name, and its device path.  Before, you would have to always provide
both a name and a path to find the requested filter, now only a path is
required.  Makes using the library a little bit less of a pain, as the
name is usually not needed.