Find and connect crossbar when directly connecting
When directly connecting the pins of the device filters to the capture
filters, it would not take in to account the potential of a crossbar
being required for the device.

This is important because most video capture devices (sans webcams)
require the use of crossbars in order to operate properly.