Add AVerMedia Stream Engine timestamp workaround
Apparently the avermedia stream engine does not handle changes in the
buffering properly (some sort of bug in the stream engine), so it'll
return bad timestamps if you try to change the buffering value from its
default.  So instead, just don't change the buffering if it detects the
stream engine being used for audio capture.