Fix possible null pointer dereferences
This appears to be happening for some users when querying audio formats
for some devices, causing a crash.  It's unknown if this is because of
bad drivers or specifically what might be happening, but it's a clear
instance of possible null pointer dereferencing so it's best to clear it
up.  This has been confirmed to fix the crash for users that experienced
it.