Don't allow 24bit audio when using default config
When specifying to use the default audio config, check to make sure that
the device does not use it if it returns 24bit audio.  24bit audio is
unusual, and OBS in particular will not support it and thus may crash.
Instead, make it try to get 16bit audio instead, which is the most
common format.

This fixes AJA devices causing a crash in OBS specifically.