Add support for more devices with uncoupled audio
Certain devices (such as the new EVGA devices, and magewell devices)
seem to use the default plug and play drivers, which do not seem to have
an audio pin.  Instead, the default drivers decouple the audio and make
the audio separate devices, forcing the user to have to select the audio
device manually in OBS.

There was already a workaround for this with elgato devices, so expand
that code to support more devices that can be "whitelisted" to this
functionality depending on their device identifier.